Moving generation of manifests files to models folder.
Removed AssociateCluster method from GitOpsDirectoryWriter
