+++ pkg/services/gitrepo/service.go
@@ -6,15 +6,16 @@ import (
 	"io/ioutil"
 	"os"
 
+	"github.com/weaveworks/weave-gitops/pkg/models"
+
 	"github.com/weaveworks/weave-gitops/pkg/git"
 	"github.com/weaveworks/weave-gitops/pkg/gitproviders"
 	"github.com/weaveworks/weave-gitops/pkg/logger"
-	"github.com/weaveworks/weave-gitops/pkg/services/automation"
 )
 
 type RepoWriter interface {
 	CreatePullRequest(ctx context.Context, info gitproviders.PullRequestInfo) error
-	WriteAndMerge(ctx context.Context, repoDir, commitMsg string, manifests []automation.Manifest) error
+	WriteAndMerge(ctx context.Context, repoDir, commitMsg string, manifests []models.Manifest) error
 	CloneRepo(ctx context.Context, branch string) (func(), string, error)
 	GetDefaultBranch(ctx context.Context) (string, error)
 	CommitAndPush(ctx context.Context, commitMsg string, filters ...func(string) bool) error
@@ -47,7 +48,7 @@ func (rw *RepoWriterSvc) CreatePullRequest(ctx context.Context, info gitprovider
 	return nil
 }
 
-func (rw *RepoWriterSvc) WriteAndMerge(ctx context.Context, repoDir, commitMsg string, manifests []automation.Manifest) error {
+func (rw *RepoWriterSvc) WriteAndMerge(ctx context.Context, repoDir, commitMsg string, manifests []models.Manifest) error {
 	for _, m := range manifests {
 		if err := rw.GitClient.Write(m.Path, m.Content); err != nil {
 			return fmt.Errorf("Failed to write manifest: %w", err)
